Exclude older nastier version of AWS api jar
There is only one version of amazon-kinesis-deaggregator available
in Maven Central, which is unfortunate because it brings in an
older version of the AWS events API, which in turn has a very bad
version range specification, causing the whole AWS internet to
be downloaded for each build.

I also made the deaggregator optional, which will help. Users that
want to include it shoudl consider doing the same exlcusion.

Fixes gh-171