Fix AbstractSpringFunctionAdapterInitializer
This initializer depends on specific function names to work properly; specifically it depends on the Function bean being called "function", Consumer bean being "consumer", and Supplier bean being "supplier" in the FunctionCatalog; however in the latest codebase it no longer works like this. In the latest codebase all types (Function, Consumer, and Supplier) get registered as "function" in FunctionalSpringApplication.  So I fixed this to only look for "function" in the catalog.

This fixes the issue and simplifies the logic, which is nice.