Adapt to restructuring changes in Spring Boot 4.0
This commit updates imports and dependencies to adapt to Spring Boot
4.0's restructuring.

FunctionEndpointInitializer has been reworked to move references to
classes that may not be on the classpath into a nested class. This
addresses a failure when SpringFactoriesLoader's reflectively
introspects the class to find its constructor.

This commit also stops using some APIs that have been deprecated
since Spring Boot 3.4.

Signed-off-by: Andy Wilkinson <andy.wilkinson@broadcom.com>